expose the build timeout in a BUILT_TIMEOUT environment variable
It would be nice to have the jenkins job timeout available during the entire build.
Tests, particularly long running tests, can make use of the timeout to evaluate and
do stuff during a test run.  Currently the work around is to set the timeout value using
the jenkins timeout plugin then using the envinject plugin to inject that value into the
job as an environment variable so that it's available globally to the job.  Although that
works it's not a great solution especially if you have tons of projects that need to be
updated with that workaround.  Also you need are using two plugins that should only
require one.  A better solution would be for the build timeout plugin to expose the
timeout value once it's set which is what this patch does.