Improve help files (wording & unused files)
Using "terminate" instead of "abort" a build because it's not any longer
only about aborting a build but there are different operations available.

Using "n" instead of "three" builds because the number of builds is
configurable.

Using "build" instead of "job" when it's about a build.

Removing help files that are not referenced anymore.