increase timeout (#136)
This plugin isn't passing with a BOM build since moving to AWS. When testing locally, the test:

`mvn clean -Dtest=AbortAndRestartOperationTest verify`

runs fine. However, if I change the timeout from 25 to 23, the local test fails.

My guess is that the AWS infra is running slower than the Azure infra, therefore causing the test to fail.

Also, it sort of makes sense that since the `once` version of the test is 25 seconds, the `twice` version of the test "should be"/"its ok" to be 50 seconds.